Rename legacy components.
