+++ assets/js/modules/analytics/components/dashboard/index.js
@@ -18,9 +18,9 @@
 
 export { default as AnalyticsAdSenseDashboardWidgetLayout } from './AnalyticsAdSenseDashboardWidgetLayout';
 export { default as AnalyticsAdSenseDashboardWidgetTopPagesTable } from './AnalyticsAdSenseDashboardWidgetTopPagesTable';
-export { default as AnalyticsDashboardWidget } from './AnalyticsDashboardWidget';
-export { default as AnalyticsDashboardWidgetOverview } from './AnalyticsDashboardWidgetOverview';
-export { default as AnalyticsDashboardWidgetSiteStats } from './AnalyticsDashboardWidgetSiteStats';
+export { default as LegacyAnalyticsDashboardWidget } from './LegacyAnalyticsDashboardWidget';
+export { default as LegacyAnalyticsDashboardWidgetOverview } from './LegacyAnalyticsDashboardWidgetOverview';
+export { default as LegacyAnalyticsDashboardWidgetSiteStats } from './LegacyAnalyticsDashboardWidgetSiteStats';
 export { default as AnalyticsDashboardWidgetTopPagesTable } from './AnalyticsDashboardWidgetTopPagesTable';
 export { default as DashboardAllTrafficWidget } from './DashboardAllTrafficWidget';
 export { default as DashboardPopularPagesWidget } from './DashboardPopularPagesWidget';
